# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Interview Platform API",
    "version": "1.0.0",
    "description": "API for AI-powered interview platform. Base URL: `/api/v1`.\n\nAuthentication: Bearer JWT (access_token). Refresh token via HttpOnly Cookie.\n\nStandard response envelope:\n- Success: `{ success: true, data, meta, request_id }`\n- Error: `{ success: false, error: { code, message, details }, request_id }`",
    "contact": {
      "name": "Khôi - Backend Core",
      "email": "khoi@example.com"
    }
  },
  "servers": [
    { "url": "/api/v1", "description": "API v1" }
  ],
  "paths": {
...
    "/companies/{company_id}/jobs": {
      "get": {
        "tags": ["Job"],
        "summary": "List jobs",
        "security": [{ "BearerAuth": [] }],
        "operationId": "jobList",
        "parameters": [
          { "$ref": "#/components/parameters/CompanyID" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "open", "paused", "closed"] } },
          { "name": "keyword", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated job list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/JobListItem" }
                    },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Job"],
        "summary": "Create job",
        "security": [{ "BearerAuth": [] }],
        "operationId": "jobCreate",
        "parameters": [{ "$ref": "#/components/parameters/CompanyID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description"],
                "properties": {
                  "title": { "type": "string", "example": "Frontend Developer" },
                  "department": { "type": "string" },
                  "level": { "type": "string", "enum": ["junior", "middle", "senior", "lead"] },
                  "location": { "type": "string" },
                  "employment_type": { "type": "string", "enum": ["full_time", "part_time", "contract", "internship"] },
                  "description": { "type": "string" },
                  "requirements": { "type": "string" },
                  "benefits": { "type": "string" },
                  "status": { "type": "string", "enum": ["draft", "open", "paused", "closed"], "default": "draft" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/JobDetail" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/companies/{company_id}/jobs/{job_id}": {
      "get": {
        "tags": ["Job"],
        "summary": "Get job detail",
        "security": [{ "BearerAuth": [] }],
        "operationId": "jobGet",
        "parameters": [
          { "$ref": "#/components/parameters/CompanyID" },
          { "$ref": "#/components/parameters/JobID" }
        ],
        "responses": {
          "200": {
            "description": "Job detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/JobDetail" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Job"],
        "summary": "Update job",
        "security": [{ "BearerAuth": [] }],
        "operationId": "jobUpdate",
        "parameters": [
          { "$ref": "#/components/parameters/CompanyID" },
          { "$ref": "#/components/parameters/JobID" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "department": { "type": "string" },
                  "level": { "type": "string" },
                  "location": { "type": "string" },
                  "employment_type": { "type": "string" },
                  "description": { "type": "string" },
                  "requirements": { "type": "string" },
                  "benefits": { "type": "string" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Job updated" } }
      },
      "delete": {
        "tags": ["Job"],
        "summary": "Soft delete job",
        "security": [{ "BearerAuth": [] }],
        "operationId": "jobDelete",
        "parameters": [
          { "$ref": "#/components/parameters/CompanyID" },
          { "$ref": "#/components/parameters/JobID" }
        ],
        "responses": { "200": { "description": "Job deleted" } }
      }
    },
    "/companies/{company_id}/jobs/{job_id}/analyze": {
      "post": {
        "tags": ["Job", "AI"],
        "summary": "Analyze job description with AI",
        "security": [{ "BearerAuth": [] }],
        "operationId": "jobAnalyze",
        "parameters": [
          { "$ref": "#/components/parameters/CompanyID" },
          { "$ref": "#/components/parameters/JobID" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "force_refresh": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JD analyzed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/JDAnalysisResult" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "CompanyID": { "name": "company_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "JobID": { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
      "PageSize": { "name": "page_size", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
    },
    "schemas": {
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
...
      "JobListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "department": { "type": "string" },
          "level": { "type": "string" },
          "status": { "type": "string" },
          "candidate_count": { "type": "integer" },
          "interview_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "JobDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "company_id": { "type": "string" },
          "title": { "type": "string" },
          "department": { "type": "string" },
          "level": { "type": "string" },
          "location": { "type": "string" },
          "employment_type": { "type": "string" },
          "description": { "type": "string" },
          "requirements": { "type": "string" },
          "benefits": { "type": "string" },
          "status": { "type": "string" },
          "ai_summary": { "type": "string", "nullable": true },
          "ai_analysis_json": { "type": "object", "nullable": true },
          "stats": { "$ref": "#/components/schemas/JobStats" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "JobStats": {
        "type": "object",
        "properties": {
          "candidate_count": { "type": "integer" },
          "interview_count": { "type": "integer" }
        }
      },
      "JDAnalysisResult": {
        "type": "object",
        "properties": {
          "summary": { "type": "string" },
          "required_skills": { "type": "array", "items": { "type": "string" } },
          "nice_to_have_skills": { "type": "array", "items": { "type": "string" } },
          "suggested_rubric": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "weight": { "type": "integer" },
                "description": { "type": "string" }
              }
            }
          },
          "suggested_questions": { "type": "array", "items": { "type": "object" } }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
}